+++ vector/src/main/java/im/vector/adapters/VectorRoomDetailsMembersAdapter.java
@@ -318,7 +318,7 @@ public class VectorRoomDetailsMembersAdapter extends BaseExpandableListAdapter {
 
                 Collection<RoomMember> activeMembers = mRoom.getActiveMembers();
                 String myUserId = mSession.getMyUserId();
-                final PowerLevels powerLevels = mRoom.getLiveState().getPowerLevels();
+                final PowerLevels powerLevels = mRoom.getState().getPowerLevels();
 
                 // search loop to extract the following members: current user, invited, administrator and others
                 for (RoomMember member : activeMembers) {
@@ -348,11 +348,11 @@ public class VectorRoomDetailsMembersAdapter extends BaseExpandableListAdapter {
                 }
 
                 // add 3rd party invite
-                Collection<RoomThirdPartyInvite> thirdPartyInvites = mRoom.getLiveState().thirdPartyInvites();
+                Collection<RoomThirdPartyInvite> thirdPartyInvites = mRoom.getState().thirdPartyInvites();
 
                 for (RoomThirdPartyInvite invite : thirdPartyInvites) {
                     // If the home server has converted the 3pid invite into a room member, do not show it
-                    if (null == mRoom.getLiveState().memberWithThirdPartyInviteToken(invite.token)) {
+                    if (null == mRoom.getState().memberWithThirdPartyInviteToken(invite.token)) {
                         ParticipantAdapterItem participant = new ParticipantAdapterItem(invite.display_name, "", null, true);
 
                         if ((!isSearchEnabled) || participant.contains(mSearchPattern)) {
@@ -754,7 +754,7 @@ public class VectorRoomDetailsMembersAdapter extends BaseExpandableListAdapter {
 
         PowerLevels powerLevels = null;
         if (null != mRoom) {
-            if (null != (powerLevels = mRoom.getLiveState().getPowerLevels())) {
+            if (null != (powerLevels = mRoom.getState().getPowerLevels())) {
                 if (powerLevels.getUserPowerLevel(participant.mUserId) >= CommonActivityUtils.UTILS_POWER_LEVEL_ADMIN) {
                     viewHolder.mMemberAvatarBadgeImageView.setVisibility(View.VISIBLE);
                     viewHolder.mMemberAvatarBadgeImageView.setImageResource(R.drawable.admin_icon);