+++ vector/src/main/java/im/vector/adapters/VectorSearchMessagesListAdapter.java
@@ -111,7 +111,7 @@ public class VectorSearchMessagesListAdapter extends VectorMessagesAdapter {
             RoomState roomState = row.getRoomState();
 
             if (null == roomState) {
-                roomState = room.getLiveState();
+                roomState = room.getState();
             }
 
             // refresh the avatar
@@ -127,7 +127,7 @@ public class VectorSearchMessagesListAdapter extends VectorMessagesAdapter {
             // display the body
             TextView bodyTextView = convertView.findViewById(R.id.messagesAdapter_body);
             // set the message text
-            EventDisplay display = new RiotEventDisplay(mContext, event, (null != room) ? room.getLiveState() : null);
+            EventDisplay display = new RiotEventDisplay(mContext, event, (null != room) ? room.getState() : null);
             CharSequence text = display.getTextualDisplay();
 
             if (null == text) {